+++ crates/c-api-macros/Cargo.toml
@@ -17,5 +17,5 @@ test = false
 doctest = false
 
 [dependencies]
-quote = "1.0"
-proc-macro2 = "1.0"
+quote = { workspace = true }
+proc-macro2 = { workspace = true }
