--- crates/misc/component-fuzz-util/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "component-fuzz-util"
-authors = ["The Wasmtime Project Developers"]
-license = "Apache-2.0 WITH LLVM-exception"
-version = "0.0.0"
-edition.workspace = true
-rust-version.workspace = true
-publish = false
-
-[lints]
-workspace = true
-
-[dependencies]
-anyhow = { workspace = true }
-arbitrary = { workspace = true, features = ["derive"] }
-proc-macro2 = "1.0"
-quote = "1.0"
-wasmtime-component-util = { workspace = true }
