--- crates/misc/component-macro-test/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "component-macro-test"
-authors = ["The Wasmtime Project Developers"]
-license = "Apache-2.0 WITH LLVM-exception"
-version = "0.0.0"
-edition.workspace = true
-rust-version.workspace = true
-publish = false
-
-[lib]
-proc-macro = true
-test = false
-doctest = false
-
-[dependencies]
-proc-macro2 = "1.0"
-quote = "1.0"
-syn = { workspace = true, features = ["full"] }
