--- crates/misc/component-test-util/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "component-test-util"
-authors = ["The Wasmtime Project Developers"]
-license = "Apache-2.0 WITH LLVM-exception"
-version = "0.0.0"
-edition.workspace = true
-rust-version.workspace = true
-publish = false
-
-[dependencies]
-env_logger = { workspace = true }
-anyhow = { workspace = true }
-arbitrary = { workspace = true, features = ["derive"] }
-wasmtime = { workspace = true, features = ["component-model", "async", "component-model-async"] }
-wasmtime-environ = { workspace = true }
-wasmtime-wast-util = { path = '../../wast-util' }
-target-lexicon = { workspace = true }
