+++ crates/test-util/src/component.rs
@@ -129,86 +129,3 @@ forward_impls! {
     Float32 => f32,
     Float64 => f64,
 }
-
-/// Helper method to apply `wast_config` to `config`.
-pub fn apply_wast_config(config: &mut Config, wast_config: &wasmtime_wast_util::WastConfig) {
-    use wasmtime_environ::TripleExt;
-    use wasmtime_wast_util::{Collector, Compiler};
-
-    config.strategy(match wast_config.compiler {
-        Compiler::CraneliftNative | Compiler::CraneliftPulley => wasmtime::Strategy::Cranelift,
-        Compiler::Winch => wasmtime::Strategy::Winch,
-    });
-    if let Compiler::CraneliftPulley = wast_config.compiler {
-        config
-            .target(&target_lexicon::Triple::pulley_host().to_string())
-            .unwrap();
-    }
-    config.collector(match wast_config.collector {
-        Collector::Auto => wasmtime::Collector::Auto,
-        Collector::Null => wasmtime::Collector::Null,
-        Collector::DeferredReferenceCounting => wasmtime::Collector::DeferredReferenceCounting,
-    });
-}
-
-/// Helper method to apply `test_config` to `config`.
-pub fn apply_test_config(config: &mut Config, test_config: &wasmtime_wast_util::TestConfig) {
-    let wasmtime_wast_util::TestConfig {
-        memory64,
-        custom_page_sizes,
-        multi_memory,
-        threads,
-        gc,
-        function_references,
-        relaxed_simd,
-        reference_types,
-        tail_call,
-        extended_const,
-        wide_arithmetic,
-        component_model_async,
-        nan_canonicalization,
-        simd,
-
-        hogs_memory: _,
-        gc_types: _,
-    } = *test_config;
-    // Note that all of these proposals/features are currently default-off to
-    // ensure that we annotate all tests accurately with what features they
-    // need, even in the future when features are stabilized.
-    let memory64 = memory64.unwrap_or(false);
-    let custom_page_sizes = custom_page_sizes.unwrap_or(false);
-    let multi_memory = multi_memory.unwrap_or(false);
-    let threads = threads.unwrap_or(false);
-    let gc = gc.unwrap_or(false);
-    let tail_call = tail_call.unwrap_or(false);
-    let extended_const = extended_const.unwrap_or(false);
-    let wide_arithmetic = wide_arithmetic.unwrap_or(false);
-    let component_model_async = component_model_async.unwrap_or(false);
-    let nan_canonicalization = nan_canonicalization.unwrap_or(false);
-    let relaxed_simd = relaxed_simd.unwrap_or(false);
-
-    // Some proposals in wasm depend on previous proposals. For example the gc
-    // proposal depends on function-references which depends on reference-types.
-    // To avoid needing to enable all of them at once implicitly enable
-    // downstream proposals once the end proposal is enabled (e.g. when enabling
-    // gc that also enables function-references and reference-types).
-    let function_references = gc || function_references.unwrap_or(false);
-    let reference_types = function_references || reference_types.unwrap_or(false);
-    let simd = relaxed_simd || simd.unwrap_or(false);
-
-    config
-        .wasm_multi_memory(multi_memory)
-        .wasm_threads(threads)
-        .wasm_memory64(memory64)
-        .wasm_function_references(function_references)
-        .wasm_gc(gc)
-        .wasm_reference_types(reference_types)
-        .wasm_relaxed_simd(relaxed_simd)
-        .wasm_simd(simd)
-        .wasm_tail_call(tail_call)
-        .wasm_custom_page_sizes(custom_page_sizes)
-        .wasm_extended_const(extended_const)
-        .wasm_wide_arithmetic(wide_arithmetic)
-        .wasm_component_model_async(component_model_async)
-        .cranelift_nan_canonicalization(nan_canonicalization);
-}
