+++ crates/versioned-export-macros/Cargo.toml
@@ -12,8 +12,8 @@ version.workspace = true
 
 [dependencies]
 syn = { workspace = true, features = ["full"] }
-quote = "1.0"
-proc-macro2 = "1.0"
+quote = { workspace = true }
+proc-macro2 = { workspace = true }
 
 [lib]
 proc-macro = true
