--- crates/wast-util/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "wasmtime-wast-util"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-publish = false
-description = "Utilities for `*.wast` files and tests in Wasmtime"
-license = "Apache-2.0 WITH LLVM-exception"
-repository = "https://github.com/bytecodealliance/wasmtime"
-
-[dependencies]
-anyhow = { workspace = true }
-serde = { workspace = true }
-serde_derive = { workspace = true }
-toml = { workspace = true }
-
-[lints]
-workspace = true
