+++ fuzz/Cargo.toml
@@ -29,17 +29,16 @@ smallvec = { workspace = true }
 wasmparser = { workspace = true }
 wasmtime = { workspace = true, features = ["winch"] }
 wasmtime-fuzzing = { workspace = true }
-component-test-util = { workspace = true }
-component-fuzz-util = { workspace = true }
+wasmtime-test-util = { workspace = true }
 log = { workspace = true }
 
 [build-dependencies]
 anyhow = { workspace = true }
-proc-macro2 = "1.0"
+proc-macro2 = { workspace = true }
 arbitrary = { workspace = true, features = ["derive"] }
 rand = { version = "0.8.0" }
-quote = "1.0"
-component-fuzz-util = { workspace = true }
+quote = { workspace = true }
+wasmtime-test-util = { workspace = true, features = ['component-fuzz'] }
 
 [features]
 default = ['fuzz-spec-interpreter']
